cambios para re entrega 2

+++ src/main/java/ar/edu/utn/dds/modelo/Division.java
@@ -1,6 +1,6 @@
 package ar.edu.utn.dds.modelo;
 
 public class Division implements Operador{
-	public int operar(int i, int j){
+	public double operar(double i, double j){
 		return i/j;
 }}
