+++ src/main/java/ar/edu/utn/dds/modelo/Multiplicacion.java
@@ -1,7 +1,7 @@
 package ar.edu.utn.dds.modelo;
 
 public class Multiplicacion implements Operador {
-	public int operar(int i, int j){
+	public double operar(double i, double j){
 		return i*j;
 	}
