+++ src/main/java/ar/edu/utn/dds/modelo/Resta.java
@@ -1,7 +1,8 @@
 package ar.edu.utn.dds.modelo;
 
 public class Resta implements Operador {
-	public int operar(int i, int j){
-		return i-j;
+	public double operar(double i, double j) {
+		return i - j;
 
-}}
+	}
+}
