+++ src/main/java/ar/edu/utn/dds/modelo/Suma.java
@@ -2,7 +2,7 @@ package ar.edu.utn.dds.modelo;
 
 public class Suma implements Operador {
 
-	public int operar(int i, int j) {
+	public double operar(double i, double j) {
 		return i + j;
 
 	}
